Updated layouts and settings state
Adds a fullscreen layout showing only the slideshow.
Adds a settings state making it possible to toggle dark/light-mode and between layouts.
The mixed layout the calendar component is moved to the slideshow space, and alternates between slides and calendar on a 15s interval, and the sl screen is expanded to ~double height.